test(input): migrate color
