+++ core/src/components/input/test/color/input.e2e.ts-snapshots/input-no-fill-color-stacked-md-ltr-Mobile-Firefox-linux.png
