+++ core/src/components/input/test/color/input.e2e.ts-snapshots/input-no-fill-color-start-md-ltr-Mobile-Safari-linux.png
