+++ core/src/components/input/test/color/input.e2e.ts-snapshots/input-outline-color-end-md-ltr-Mobile-Chrome-linux.png
