+++ core/src/components/input/test/color/input.e2e.ts-snapshots/input-outline-color-fixed-md-ltr-Mobile-Firefox-linux.png
