+++ core/src/components/input/test/color/input.e2e.ts-snapshots/input-outline-color-floating-md-ltr-Mobile-Safari-linux.png
