+++ core/src/components/input/test/color/input.e2e.ts-snapshots/input-outline-color-stacked-md-ltr-Mobile-Chrome-linux.png
