+++ core/src/components/input/test/color/input.e2e.ts-snapshots/input-solid-color-end-md-ltr-Mobile-Chrome-linux.png
