+++ core/src/components/input/test/color/input.e2e.ts-snapshots/input-solid-color-fixed-md-ltr-Mobile-Chrome-linux.png
