+++ core/src/components/input/test/color/input.e2e.ts-snapshots/input-solid-color-stacked-md-ltr-Mobile-Firefox-linux.png
